requirements.txt
Jupyter
black==22.12.0
db-sqlite3==0.0.1
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.3
plotly_express==0.4.1
requests==2.26.0
seaborn==0.11.2
SQLAlchemy==1.4.45
nbformat==5.7.3
pytest==7.2.1











